:truck: Adding the pages folder to src
Partially changing the folder structure to make the App file hight-level to
avoid the error "React Hooks must be called in a React function component"
